* Use of ptrdiff_t whenever necessary (IA-64 port) ; * Changed behaviour of preparsing code, faster startup ; * Fixed DecodePSM(), cannot test ; * Cleaned up ParsePES(), fixed a bug which could reject very small but   valid PES packets ; * Fixed cosmetic bugs in vpar_synchro frame_rate display.

+++ include/input_ext-dec.h
@@ -2,7 +2,7 @@
  * input_ext-dec.h: structures exported to the VideoLAN decoders
  *****************************************************************************
  * Copyright (C) 1999, 2000 VideoLAN
- * $Id: input_ext-dec.h,v 1.5 2000/12/22 17:53:30 massiot Exp $
+ * $Id: input_ext-dec.h,v 1.6 2000/12/26 19:14:46 massiot Exp $
  *
  * Authors:
  *
@@ -400,7 +400,7 @@ static __inline__ void RealignBits( bit_stream_t * p_bit_stream )
 static __inline__ void GetChunk( bit_stream_t * p_bit_stream,
                                  byte_t * p_buffer, size_t i_buf_len )
 {
-    int     i_available;
+    ptrdiff_t           i_available;
 
     if( (i_available = p_bit_stream->p_end - p_bit_stream->p_byte)
             >= i_buf_len )
